@@ -1,13 +1,11 @@
 // src/components/ClaudeAssistant.tsx
-import React, { useState, useRef, useEffect, useCallback } from 'react';
 import { BoltIcon, MessageCircleIcon, SearchIcon } from 'lucide-react';
-
+import React, { useState, useRef, useEffect, useCallback } from 'react';
 import AccessibleTabs from './AccessibleTabs';
-import MessageBubble from './MessageBubble';
 import { StatusBadge, LoadingIndicator, TypingIndicator } from './Claude';
-
-import { useToast } from '@/context/ToastContext';
+import MessageBubble from './MessageBubble';
 import { useClaude } from '@/context/ClaudeProvider';
+import { useToast } from '@/context/ToastContext';
 
 interface ClaudeAssistantProps {
   selectedText?: string;
