@@ -1,9 +1,8 @@
 // src/hooks/useCommands.ts
 import { useMemo } from 'react';
-
 import { useAppContext, View } from '@/context/AppContext';
-import { Command } from '@/types/commands';
 import { useToast } from '@/context/ToastContext';
+import { Command } from '@/types/commands';
 
 export function useCommands(
   selectedText?: string,
