@@ -1,10 +1,9 @@
 // src/components/ExportDialog.tsx
-import React, { useState } from 'react';
 import { X, Download, FileText, File, FileImage, Settings } from 'lucide-react';
-
-import { ExportFormat } from '@/types/writing';
-import { exportService, ExportOptions } from '@/services/exportService';
+import React, { useState } from 'react';
 import { useToast } from '@/context/ToastContext';
+import { exportService, ExportOptions } from '@/services/exportService';
+import { ExportFormat } from '@/types/writing';
 
 interface ExportDialogProps {
   isOpen: boolean;
