@@ -1,9 +1,8 @@
 // src/components/Panels/AnalysisPanel.tsx - Upgraded Version
-import React, { useEffect, useState } from 'react';
 import { BarChart3, TrendingUp, Clock, Target } from 'lucide-react';
-
-import { useAppContext } from '@/context/AppContext';
+import React, { useEffect, useState } from 'react';
 import WritingAnalyticsView from '@/components/Analytics/WritingAnalyticsView';
+import { useAppContext } from '@/context/AppContext';
 
 interface WritingSession {
   date: string;
