@@ -1,13 +1,12 @@
 // File: src/components/Panels/SettingsPanel.tsx - Enhanced Claude Setup
-import React, { useState, useEffect } from 'react';
 import { Eye, EyeOff, ExternalLink, CheckCircle, AlertCircle, Info } from 'lucide-react';
-
-import { useAppContext } from '@/context/AppContext';
-import { useToast } from '@/context/ToastContext';
-import claudeService from '@/services/claudeService';
+import React, { useState, useEffect } from 'react';
 import BackupControls from '@/components/Settings/BackupControls';
 import SnapshotHistoryDialog from '@/components/Settings/SnapshotHistoryDialog';
 import { Button } from '@/components/ui/Button';
+import { useAppContext } from '@/context/AppContext';
+import { useToast } from '@/context/ToastContext';
+import claudeService from '@/services/claudeService';
 
 const SettingsPanel: React.FC = () => {
   const { state, claudeActions, claude } = useAppContext();
