@@ -1,12 +1,11 @@
 // File: src/components/Settings/SnapshotHistoryDialog.tsx
-import React, { useEffect, useMemo, useState } from 'react';
 import { formatDistanceToNow } from 'date-fns';
-
-import { Dialog } from '@/components/ui/dialog';
+import React, { useEffect, useMemo, useState } from 'react';
 import { Button } from '@/components/ui/Button';
+import { Dialog } from '@/components/ui/dialog';
+import { ScrollArea } from '@/components/ui/scroll-area';
 import { useCurrentProject } from '@/context/AppContext';
 import snapshotService, { SnapshotMeta } from '@/services/snapshotAdapter';
-import { ScrollArea } from '@/components/ui/scroll-area';
 
 interface SnapshotHistoryDialogProps {
   open: boolean;
