@@ -1,9 +1,8 @@
 import React, { useCallback } from 'react';
-
-import { useFocusMode } from '@/hooks/useFocusMode';
 import { useAppContext } from '@/context/AppContext';
-import snapshotService from '@/services/snapshotAdapter';
 import { useCurrentProject } from '@/context/AppContext';
+import { useFocusMode } from '@/hooks/useFocusMode';
+import snapshotService from '@/services/snapshotAdapter';
 
 export type TopbarProps = {
   onOpenNotifications?: () => void;
