@@ -1,14 +1,13 @@
 // src/components/Writing/TipTapEditor.tsx
-import React, { useEffect, useRef } from 'react';
-import { EditorContent, useEditor } from '@tiptap/react';
-import StarterKit from '@tiptap/starter-kit';
-import Placeholder from '@tiptap/extension-placeholder';
-import Typography from '@tiptap/extension-typography';
 import CharacterCount from '@tiptap/extension-character-count';
 import History from '@tiptap/extension-history';
-
-import { cn } from '@/utils/cn';
+import Placeholder from '@tiptap/extension-placeholder';
+import Typography from '@tiptap/extension-typography';
+import { EditorContent, useEditor } from '@tiptap/react';
+import StarterKit from '@tiptap/starter-kit';
+import React, { useEffect, useRef } from 'react';
 import { useAdvancedFocusMode } from '@/hooks/useAdvancedFocusMode';
+import { cn } from '@/utils/cn';
 
 type Props = {
   value: string;
