@@ -1,7 +1,6 @@
 // src/components/ui/ToastMessage.tsx
-import React from 'react';
 import { AlertCircle, CheckCircle, Info, XCircle } from 'lucide-react';
-
+import React from 'react';
 import { cn } from '@/utils/cn';
 
 type ToastType = 'success' | 'error' | 'info' | 'warning';
